Applying admin policy

+++ src/Umbraco.Cms.Api.Management/Controllers/User/BulkDeleteUsersController.cs
@@ -1,21 +1,28 @@
 ﻿using Asp.Versioning;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Umbraco.Cms.Api.Management.ViewModels.User;
 using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
+using Umbraco.Cms.Web.Common.Authorization;
 
 namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 [ApiVersion("1.0")]
 public class BulkDeleteUsersController : UserControllerBase
 {
+    private readonly IAuthorizationService _authorizationService;
     private readonly IUserService _userService;
     private readonly IBackOfficeSecurityAccessor _backOfficeSecurityAccessor;
 
-    public BulkDeleteUsersController(IUserService userService, IBackOfficeSecurityAccessor backOfficeSecurityAccessor)
+    public BulkDeleteUsersController(
+        IAuthorizationService authorizationService,
+        IUserService userService,
+        IBackOfficeSecurityAccessor backOfficeSecurityAccessor)
     {
+        _authorizationService = authorizationService;
         _userService = userService;
         _backOfficeSecurityAccessor = backOfficeSecurityAccessor;
     }
@@ -26,6 +33,14 @@ public class BulkDeleteUsersController : UserControllerBase
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> DeleteUsers(DeleteUsersRequestModel model)
     {
+        AuthorizationResult authorizationResult = await _authorizationService.AuthorizeAsync(User, model.UserIds,
+            $"New{AuthorizationPolicies.AdminUserEditsRequireAdmin}");
+
+        if (!authorizationResult.Succeeded)
+        {
+            return Forbidden();
+        }
+
         UserOperationStatus result = await _userService.DeleteAsync(CurrentUserKey(_backOfficeSecurityAccessor), model.UserIds);
 
         return result is UserOperationStatus.Success
